refactor: Move logging package into utils folder
